move all reducer code to the index, no more combinereducers, to modify state shape
